+++ backend-services/voting-app/src/main/java/org/cardano/foundation/voting/service/leader_board/DefaultLeaderBoardService.java
@@ -7,11 +7,10 @@ import org.cardano.foundation.voting.client.ChainFollowerClient;
 import org.cardano.foundation.voting.domain.Leaderboard;
 import org.cardano.foundation.voting.repository.VoteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.zalando.problem.Problem;
 
-import java.util.Arrays;
 import java.util.Optional;
 
 import static java.util.stream.Collectors.toMap;
@@ -27,8 +26,45 @@ public class DefaultLeaderBoardService implements LeaderBoardService {
     @Autowired
     private VoteRepository voteRepository;
 
-    @Autowired
-    private Environment env;
+    @Value("${leaderboard.force.results:false}")
+    private boolean forceLeaderboardResultsAvailability;
+
+    @Override
+    public Either<Problem, Boolean> isEventLeaderboardAvailable(String event) {
+        var eventDetailsE = chainFollowerClient.getEventDetails(event);
+        if (eventDetailsE.isEmpty()) {
+            return Either.left(Problem.builder()
+                    .withTitle("ERROR_GETTING_EVENT_DETAILS")
+                    .withDetail("Unable to get event details from chain-tip follower service, event:" + event)
+                    .withStatus(INTERNAL_SERVER_ERROR)
+                    .build()
+            );
+        }
+        var maybeEventDetails = eventDetailsE.get();
+        if (maybeEventDetails.isEmpty()) {
+            return Either.left(Problem.builder()
+                    .withTitle("UNRECOGNISED_EVENT")
+                    .withDetail("Unrecognised event, event:" + event)
+                    .withStatus(BAD_REQUEST)
+                    .build()
+            );
+        }
+        var eventDetails = maybeEventDetails.orElseThrow();
+
+        return isEventLeaderboardAvailable(eventDetails);
+    }
+
+    private Either<Problem, Boolean> isEventLeaderboardAvailable(ChainFollowerClient.EventDetailsResponse eventDetails) {
+        if (forceLeaderboardResultsAvailability) {
+            return Either.right(true);
+        }
+
+        if (eventDetails.highLevelResultsWhileVoting()) {
+            return Either.right(true);
+        }
+
+        return Either.right(eventDetails.finished());
+    }
 
     @Override
     public Either<Problem, Leaderboard.ByEvent> getEventLeaderboard(String event) {
@@ -52,7 +88,13 @@ public class DefaultLeaderBoardService implements LeaderBoardService {
         }
         var eventDetails = maybeEventDetails.orElseThrow();
 
-        if (!isDevEnv() && eventDetails.finished() && eventDetails.highLevelResultsWhileVoting()) {
+        var eventLeaderboardAvailableE = isEventLeaderboardAvailable(eventDetails);
+        if (eventLeaderboardAvailableE.isEmpty()) {
+            return Either.left(eventLeaderboardAvailableE.getLeft());
+        }
+        var isEventLeaderBoardAvailable = eventLeaderboardAvailableE.get();
+
+        if (!isEventLeaderBoardAvailable) {
             return Either.left(Problem.builder()
                     .withTitle("VOTING_RESULTS_NOT_AVAILABLE")
                     .withDetail("High level voting results not available until voting event finishes.")
@@ -116,7 +158,13 @@ public class DefaultLeaderBoardService implements LeaderBoardService {
         }
         var categoryDetails = maybeCategory.orElseThrow();
 
-        if (!isDevEnv() && eventDetails.finished() && eventDetails.categoryResultsWhileVoting()) {
+        var categoryLeaderboardAvailableE = isCategoryLeaderboardAvailable(eventDetails, categoryDetails);
+        if (categoryLeaderboardAvailableE.isEmpty()) {
+            return Either.left(categoryLeaderboardAvailableE.getLeft());
+        }
+
+        var isCategoryLeaderBoardAvailable = categoryLeaderboardAvailableE.get();
+        if (!isCategoryLeaderBoardAvailable) {
             return Either.left(Problem.builder()
                     .withTitle("VOTING_RESULTS_NOT_AVAILABLE")
                     .withDetail("High level voting results not available until voting event finishes.")
@@ -142,8 +190,54 @@ public class DefaultLeaderBoardService implements LeaderBoardService {
         );
     }
 
-    private boolean isDevEnv() {
-        return Arrays.stream(env.getActiveProfiles()).anyMatch(env -> env.contains("dev"));
+    @Override
+    public Either<Problem, Boolean> isCategoryLeaderboardAvailable(String event, String category) {
+        var eventDetailsE = chainFollowerClient.getEventDetails(event);
+        if (eventDetailsE.isEmpty()) {
+            return Either.left(Problem.builder()
+                    .withTitle("ERROR_GETTING_EVENT_DETAILS")
+                    .withDetail("Unable to get event details from chain-tip follower service, event:" + event)
+                    .withStatus(INTERNAL_SERVER_ERROR)
+                    .build()
+            );
+        }
+        var maybeEventDetails = eventDetailsE.get();
+        if (maybeEventDetails.isEmpty()) {
+            return Either.left(Problem.builder()
+                    .withTitle("UNRECOGNISED_EVENT")
+                    .withDetail("Unrecognised event, event:" + event)
+                    .withStatus(BAD_REQUEST)
+                    .build()
+            );
+        }
+        var eventDetails = maybeEventDetails.orElseThrow();
+
+        var maybeCategory = eventDetails.categoryDetailsById(category);
+
+        if (maybeCategory.isEmpty()) {
+            return Either.left(Problem.builder()
+                    .withTitle("UNRECOGNISED_CATEGORY")
+                    .withDetail("Unrecognised category, category:" + category)
+                    .withStatus(BAD_REQUEST)
+                    .build()
+            );
+        }
+        var categoryDetails = maybeCategory.orElseThrow();
+
+        return isCategoryLeaderboardAvailable(eventDetails, categoryDetails);
+    }
+
+    private Either<Problem, Boolean> isCategoryLeaderboardAvailable(ChainFollowerClient.EventDetailsResponse eventDetails,
+                                                                    ChainFollowerClient.CategoryDetailsResponse categoryDetails) {
+        if (forceLeaderboardResultsAvailability) {
+            return Either.right(true);
+        }
+
+        if (eventDetails.categoryResultsWhileVoting()) {
+            return Either.right(true);
+        }
+
+        return Either.right(eventDetails.finished());
     }
 
 }