+++ ui/cip-1694/src/components/common/Footer/__tests__/Footer.test.tsx
@@ -1,8 +1,7 @@
-
 import React from 'react';
 import '@testing-library/jest-dom';
 import { expect } from '@jest/globals';
-import { screen, within, waitFor, cleanup,render } from '@testing-library/react';
+import { screen, within, waitFor, cleanup, render } from '@testing-library/react';
 import { Footer } from '../Footer';
 
 describe('Footer', () => {
@@ -24,10 +23,38 @@ describe('Footer', () => {
         `© ${new Date().getFullYear()} CIP-1694 Ratification. All rights reserved.`
       );
     });
-
-    // add tests for t-and-c
-    // add tests for privacy
-    // add tests for status
-    // add tests for discord
   });
+
+  test('should display proper t&c', async () => {
+    render(<Footer />);
+
+    await waitFor(async () => {
+      const footer = await screen.queryByTestId('footer');
+      expect(footer).not.toBeNull();
+
+      const tandc = await within(footer).queryByTestId('t-and-c');
+      expect(tandc).not.toBeNull();
+      expect(tandc.textContent).toEqual('Terms & Conditions');
+      expect(tandc.attributes.getNamedItem('href').value).toEqual('pdf');
+      expect(tandc.attributes.getNamedItem('type').value).toEqual('application/pdf');
+    });
+  });
+
+  test('should display proper privacy policy', async () => {
+    render(<Footer />);
+
+    await waitFor(async () => {
+      const footer = await screen.queryByTestId('footer');
+      expect(footer).not.toBeNull();
+
+      const privacy = await within(footer).queryByTestId('privacy');
+      expect(privacy).not.toBeNull();
+      expect(privacy.textContent).toEqual('Privacy');
+      expect(privacy.attributes.getNamedItem('href').value).toEqual('pdf');
+      expect(privacy.attributes.getNamedItem('type').value).toEqual('application/pdf');
+    });
+  });
+
+  // add tests for status
+  // add tests for discord
 });